Dont assume nsenter, docker-enter in same folder
In boot2docker, nsenter is currently not on the PATH.
It has to be specifically located in the folder where
`docker-enter` is located.

For a standard distribution nsenter is on the PATH.
In coreos, `nsenter` is built-in and installed in the read-only
partition as `/usr/bin/nsenter`.
A systemd service must not installs docker-enter in `/usr/bin`.

This change makes sure that we force the full path of nsenter if and
only if it exists in the same directory.